+++ examples/advanced/main.tf
@@ -28,6 +28,13 @@ module "firewall_rules" {
   stage            = "prd"
   default_location = local.location
 
-  ipg_application_lz_id = azurerm_ip_group.application_lz.id
-  ipg_platform_id       = azurerm_ip_group.platform.id
+  ipg_dnsprivateresolver_id = azurerm_ip_group.dnsprivateresolver.id
+  ipg_azure_dc_id           = azurerm_ip_group.azure_dc.id
+  ipg_application_lz_id     = azurerm_ip_group.application_lz.id
+  ipg_platform_id           = azurerm_ip_group.platform.id
+  bastion_config = {
+    ipg_bastion_id     = azurerm_ip_group.bastion.id
+    ipg_rdp_access_ids = [azurerm_ip_group.application_lz.id]
+    ipg_ssh_access_ids = [azurerm_ip_group.application_lz.id]
+  }
 }