+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.2.0] - 2024-04-08
+
+### Added
+- platform rules for bastion
+
+### Changed
+- made rules for private dns resolver and azure dcs optional
+- changed internal ordering of rules
+
+### Removed
+
+### Fixed
+
+
 ## [1.1.0] - 2023-11-17
 
 ### Added